+++ Vkm.Api.Core/Basic/IconSize.cs
@@ -5,11 +5,11 @@ namespace Vkm.Api.Basic
     [Serializable]
     public struct IconSize
     {
-        public readonly int Width;
-        public readonly int Height;
+        public readonly ushort Width;
+        public readonly ushort Height;
 
 
-        public IconSize(int width, int height)
+        public IconSize(ushort width, ushort height)
         {
             Width = width;
             Height = height;
